requirements.txt
streamlit
pandas
numpy
plotly
pymongo
scikit-learn
joblib
python-dotenv
